requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
faiss-cpu>=1.7.4
psycopg2-binary>=2.9.9
httpx>=0.26.0
pydantic>=2.5.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
websockets>=12.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
